feat(forms): Implement a prototype of typed forms.
This PR adds generics to `AbstractControl` classes as well as `FormBuilder`. The old behavior of these classes can be obtained by providing `<any>` where a generic is expected. There are still two major bugs: types are not inferred on FormArray, and FormState shows up in control types unexpectedly.

BREAKING CHANGE: This adds generics to several forms APIs, and will break any uses of the modified APIs that have not been migrated to provide an explicit `<any>`.

Specifically, `AbstractControl` and descendents (`FormControl`, `FormGroup`, and `FormArray`) have been modified, as well as `FormBuilder`.

Closes #13721.